Added function to handle form information, and render city's from my database to the screen, along with the current temperature that is retreived from openweathermaps api.
Added function to handle if user inputs a state name into the form. Open weather API end points can take in a state as well, but not as an abbrievation so my function will traspose it to the state's full name. Adding state to the end point will help ensure the correct city data is retrieved.